feat(new schema): allow public access to EbeanLocalAccess::readSqlRows
