+++ dao-impl/ebean-dao/src/main/java/com/linkedin/metadata/dao/utils/EBeanDAOUtils.java
@@ -1,15 +1,23 @@
 package com.linkedin.metadata.dao.utils;
 
+import com.linkedin.metadata.aspect.AuditedAspect;
 import com.linkedin.metadata.dao.EbeanMetadataAspect;
 import com.linkedin.metadata.dao.ListResult;
 import com.linkedin.data.template.RecordTemplate;
 import com.linkedin.metadata.aspect.SoftDeletedAspect;
+import io.ebean.SqlRow;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.sql.Timestamp;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import lombok.extern.slf4j.Slf4j;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+import org.json.simple.parser.ParseException;
 
 
 /**
@@ -23,6 +31,7 @@ public class EBeanDAOUtils {
   // String stored in metadata_aspect table for soft deleted aspect
   private static final RecordTemplate DELETED_METADATA = new SoftDeletedAspect().setGma_deleted(true);
   public static final String DELETED_VALUE = RecordUtils.toJsonString(DELETED_METADATA);
+  private static final long LATEST_VERSION = 0L;
 
   private EBeanDAOUtils() {
     // Utils class
@@ -118,5 +127,53 @@ public class EBeanDAOUtils {
     // Convert metadata string to record template object
     final RecordTemplate metadataRecord = RecordUtils.toRecordTemplate(aspectClass, aspect.getMetadata());
     return metadataRecord.equals(DELETED_METADATA);
+  }
+
+
+  /**
+   * Read {@link SqlRow} list into a {@link EbeanMetadataAspect} list.
+   * @param sqlRows list of {@link SqlRow}
+   * @return list of {@link EbeanMetadataAspect}
+   */
+  // TODO: make this method private once all users' custom SQL queries have been replaced by DAO-supported methods
+  public static List<EbeanMetadataAspect> readSqlRows(List<SqlRow> sqlRows) {
+    return sqlRows.stream().flatMap(sqlRow -> {
+      List<String> columns = new ArrayList<>();
+      sqlRow.keySet().stream().filter(key -> key.startsWith(SQLSchemaUtils.ASPECT_PREFIX) && sqlRow.get(key) != null).forEach(columns::add);
+
+      return columns.stream().map(columnName -> {
+        EbeanMetadataAspect ebeanMetadataAspect = new EbeanMetadataAspect();
+        String urn = sqlRow.getString("urn");
+        AuditedAspect auditedAspect = RecordUtils.toRecordTemplate(AuditedAspect.class, sqlRow.getString(columnName));
+        EbeanMetadataAspect.PrimaryKey primaryKey = new EbeanMetadataAspect.PrimaryKey(urn, auditedAspect.getCanonicalName(), LATEST_VERSION);
+        ebeanMetadataAspect.setKey(primaryKey);
+        ebeanMetadataAspect.setCreatedBy(auditedAspect.getLastmodifiedby());
+        ebeanMetadataAspect.setCreatedOn(Timestamp.valueOf(auditedAspect.getLastmodifiedon()));
+        ebeanMetadataAspect.setCreatedFor(auditedAspect.getCreatedfor());
+        ebeanMetadataAspect.setMetadata(extractAspectJsonString(sqlRow.getString(columnName)));
+        return ebeanMetadataAspect;
+      });
+    }).collect(Collectors.toList());
+  }
+
+  /**
+   * Extract aspect json string from an AuditedAspect string in its DB format. Return null if aspect json string does not exist.
+   * @param auditedAspect an AuditedAspect string in its DB format
+   * @return A string which can be deserialized into Aspect object.
+   */
+  @Nullable
+  private static String extractAspectJsonString(@Nonnull final String auditedAspect) {
+    try {
+      JSONParser jsonParser = new JSONParser();
+      JSONObject map = (JSONObject) jsonParser.parse(auditedAspect);
+      if (map.containsKey("aspect")) {
+        return map.get("aspect").toString();
+      }
+
+      return null;
+    } catch (ParseException parseException) {
+      log.error(String.format("Failed to parse string %s as AuditedAspect. Exception: %s", auditedAspect, parseException));
+      throw new RuntimeException(parseException);
+    }
   }
 }
